pyproject.toml
[tool.poetry]
name = "darwin"
version = "0.1.0"
description = ""
authors = ["Clement Grimal <clement.grimal@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6.2"
requests = "^2.25.1"
lxml = "^4.6.3"
pyquery = "^1.4.3"
dateparser = "^1.0.0"
python-slugify = "^5.0.0"
mutagen = "^1.45.1"

[tool.poetry.dev-dependencies]
black = "^21.4b2"
isort = "^5.8.0"
flake8 = "^3.9.1"

[build-system]
requires = ["poetry-core>=1.0.0", "pip>=21.1.1"]
build-backend = "poetry.core.masonry.api"
